v1.1.1

+++ src/DS1624.cpp
@@ -143,7 +143,7 @@ int DS1624::readMemory( uint8_t startAddr, int count, uint8_t* data ) {
     if ( Wire.endTransmission( false ) != 0 ) // Do not send STOP signal
       return 0;
 
-    uint8_t bytesRead = Wire.requestFrom( (int)_address, (int)step, true );
+    uint8_t bytesRead = Wire.requestFrom( _address, (size_t)step, true );
     if ( bytesRead > 0 ) {
       address += bytesRead;
       remaining -= bytesRead;
